Refactor: new utils.py file to factorize code.
